pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lemur-archive"
version = "0.1.0"
description = "Lemur X-ray galaxy cluster archive pipeline and API"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE.md" }
keywords = ["astronomy", "xray", "fastapi", "pipeline"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
authors = [
  { name = "Carter Rhea" }
]
urls = { Repository = "https://github.com/crhea93/Lemur", Issues = "https://github.com/crhea93/Lemur/issues" }
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.29",
]

[project.optional-dependencies]
pipeline = [
  "mysql-connector-python>=8.0",
  "numpy>=1.24",
]
test = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "httpx>=0.27",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "httpx>=0.27",
  "ruff>=0.6.0",
  "black>=24.0.0",
  "mypy>=1.11",
]

[project.scripts]
lemur-pipeline = "Pipeline.pipeline:main"
lemur-ingest-sql = "api.ingest_sql_dump:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["Pipeline*", "api*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
files = ["api", "Pipeline/config.py", "Pipeline/pipeline.py", "tests"]
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
